On workers, provide a mapping between the mount point labels and the local fields

+++ cm/util/worker.py
@@ -180,28 +180,11 @@ class ConsoleManager(BaseConsoleManager):
                 raise Exception("Mount point parsing failure.")
         except Exception, e:
             log.error("Error mounting devices: {0}\n Attempting to continue, but failure likely...".format(e))
-#             # malformed json, revert to old behavior.
-#             log.info("Mounting NFS directories from master with IP address: %s..." %
-#                      master_ip)
-#             # Build list of mounts based on cluster type
-#             if self.cluster_type == 'Galaxy':
-#                 mount_points.append(
-#                     ('nfs_tools', self.app.path_resolver.galaxy_tools, 'nfs', master_ip, ''))
-#                 mount_points.append(
-#                     ('nfs_indices', self.app.path_resolver.galaxy_indices, 'nfs', master_ip, ''))
-#             if self.cluster_type == 'Galaxy' or self.cluster_type == 'Data':
-#                 mount_points.append(
-#                     ('nfs_data', self.app.path_resolver.galaxy_data, 'nfs', master_ip, ''))
-#             # Mount master's transient storage regardless of cluster type
-#             mount_points.append(('nfs_tfs', '/mnt/transient_nfs', 'nfs', master_ip, ''))
         # Mount SGE regardless of cluster type
         mount_points.append(('nfs_sge', self.app.path_resolver.sge_root, 'nfs', master_ip, ''))
 
-        # <KWS>Mount Hadoop regardless of cluster type
+        # Mount Hadoop regardless of cluster type
         mount_points.append(('nfs_hadoop', paths.P_HADOOP_HOME, 'nfs', master_ip, ''))
-
-        # Mount master's transient storage regardless of cluster type
-        # mount_points.append(('nfs_tfs', '/mnt/transient_nfs'))
 
         for i, extra_mount in enumerate(self._get_extra_nfs_mounts()):
             mount_points.append(('extra_mount_%d' % i, extra_mount, 'nfs', master_ip, ''))
@@ -214,7 +197,15 @@ class ConsoleManager(BaseConsoleManager):
                 continue
             ret_code = self.mount_disk(fs_type, server, path, mount_options)
             status = 1 if ret_code == 0 else -1
-            setattr(self, label, status)
+            # Provide a mapping between the mount point labels and the local fields
+            # Given tools & data file systems have been merged, this mapping does
+            # not distinguish bewteen those but simply chooses the data field.
+            labels_to_fields = {
+                'galaxy': 'nfs_data',
+                'galaxyIndices': 'nfs_indices',
+                'transient_nfs': 'nfs_tfs'
+            }
+            setattr(self, labels_to_fields.get(label, label), status)
         # Filter out any differences between new and old mount points and unmount
         # the extra ones
         umount_points = [ump for ump in self.mount_points if ump not in mount_points]
@@ -334,8 +325,12 @@ class ConsoleManager(BaseConsoleManager):
     # Updating etc host by fetching the master's etc/hosts file
     # # this is necessary for hadoop ssh component
     def sync_etc_host(self, sync_path=paths.P_ETC_TRANSIENT_PATH):
-
-        shutil.copyfile(sync_path, "/etc/hosts")
+        if os.path.exists(sync_path):
+            log.debug("Synced /etc/hosts with %s" % sync_path)
+            shutil.copyfile(sync_path, "/etc/hosts")
+        else:
+            log.warning("Sync path %s not available; cannot sync /etc/hosts"
+                % sync_path)
 
     def _get_extra_nfs_mounts(self):
         return self.app.ud.get('extra_nfs_mounts', [])
